+++ www/components/restangular/test/restangularSpec.js
@@ -4,25 +4,7 @@ describe("Restangular", function() {
   var accountsModel, restangularAccounts, restangularAccount0, restangularAccount1;
   var messages, newAccount;
 
-  // Utils
-  // Apply "sanitizeRestangularOne" function to an array of items
-  function sanitizeRestangularAll(items) {
-    var all = _.map(items, function(item) {
-      return sanitizeRestangularOne(item);
-    });
-    return sanitizeRestangularOne(all);
-  };
-
-  // Remove all Restangular/AngularJS added methods in order to use Jasmine toEqual between the retrieve resource and the model
-  function sanitizeRestangularOne(item) {
-    return _.omit(item, "route", "parentResource", "getList", "get", "post", "put", "remove", "head", "trace", "options", "patch",
-      "$then", "$resolved", "restangularCollection", "customOperation", "customGET", "customPOST",
-      "customPUT", "customDELETE", "customGETLIST", "$getList", "$resolved", "restangularCollection", "one", "all","doGET", "doPOST",
-      "doPUT", "doDELETE", "doGETLIST", "addRestangularMethod", "getRestangularUrl");
-  };
-
   // Load required modules
-  beforeEach(angular.mock.module("ngResource"));
   beforeEach(angular.mock.module("restangular"));
 
   // Init HTTP mock backend and Restangular resources
@@ -45,13 +27,19 @@ describe("Restangular", function() {
     $httpBackend.when("OPTIONS", "/accounts").respond();
 
     $httpBackend.whenGET("/accounts").respond(accountsModel);
+    $httpBackend.whenJSONP("/accounts").respond(accountsModel);
+    $httpBackend.whenGET("/accounts/0,1").respond(accountsModel);
     $httpBackend.whenGET("/accounts/messages").respond(messages);
     $httpBackend.whenGET("/accounts/1/message").respond(messages[0]);
     $httpBackend.whenGET("/accounts/1/messages").respond(messages);
-    $httpBackend.whenGET("/accounts/0").respond(accountsModel[1]);
+    $httpBackend.whenGET("/accounts/0").respond(accountsModel[0]);
     $httpBackend.whenGET("/accounts/1").respond(accountsModel[1]);
+    $httpBackend.whenJSONP("/accounts/1").respond(accountsModel[1]);
     $httpBackend.whenGET("/accounts/1/transactions").respond(accountsModel[1].transactions);
     $httpBackend.whenGET("/accounts/1/transactions/1").respond(accountsModel[1].transactions[1]);
+
+    // Full URL
+    $httpBackend.whenGET('http://accounts.com/all').respond(accountsModel);
 
     $httpBackend.whenPOST("/accounts").respond(function(method, url, data, headers) {
       var newData = angular.fromJson(data);
@@ -87,10 +75,216 @@ describe("Restangular", function() {
     $httpBackend.verifyNoOutstandingRequest();
   });
 
+  describe("Interceptors", function() {
+    it("Should add multiple request and response interceptors", function() {
+      Restangular.addRequestInterceptor(function(elem) {
+        elem.firstRequestInterceptor = true;
+        return elem;
+      });
+      Restangular.addRequestInterceptor(function(elem) {
+        elem.secondRequestInterceptor = true;
+        return elem;
+      });
+      Restangular.addFullRequestInterceptor(function(elem) {
+        elem.thirdRequestInterceptor = true;
+        return {
+          element: elem
+        };
+      });
+
+      Restangular.addResponseInterceptor(function(elem) {
+        elem.firstResponseInterceptor = true;
+        return elem;
+      });
+
+      Restangular.addResponseInterceptor(function(elem) {
+        elem.secondResponseInterceptor = true;
+        return elem;
+      });
+
+      $httpBackend.whenPOST("/list").respond(function(method, url, data, headers) {
+        var elem = angular.fromJson(data);
+        expect(elem.firstRequestInterceptor).toBeDefined();
+        expect(elem.secondRequestInterceptor).toBeDefined();
+        expect(elem.thirdRequestInterceptor).toBeDefined();
+        return [200, elem, ""];
+      });
+
+      $httpBackend.expectPOST('/list');
+
+       Restangular.all('list').post({name: "Gonto"}).then(function(elem) {
+        expect(elem.firstResponseInterceptor).toBeDefined();
+        expect(elem.secondResponseInterceptor).toBeDefined();
+       });
+
+       $httpBackend.flush();
+
+
+    });
+  });
+
+  describe("Transformers", function() {
+    it("Should decorate element both on server and local by default", function() {
+
+      Restangular.extendModel('accounts', function(account) {
+        account.extended = function() {return true;}
+        return account;
+      });
+
+      Restangular.one('accounts', 1).get().then(function(account) {
+        expect(account.extended).toBeDefined();
+      });
+
+      var local = {};
+      Restangular.restangularizeElement(null, local, 'accounts');
+      expect(local.extended).toBeDefined();
+
+      $httpBackend.flush();
+    });
+
+    it("Should decorate element only server if config set", function() {
+
+      Restangular.extendModel('accounts', function(account) {
+        account.extended = function() {return true;}
+        return account;
+      });
+
+      Restangular.setTransformOnlyServerElements(true)
+
+      Restangular.one('accounts', 1).get().then(function(account) {
+        expect(account.extended).toBeDefined();
+      });
+
+      var local = {};
+      Restangular.restangularizeElement(null, local, 'accounts');
+      expect(local.extended).toBeUndefined();
+
+      $httpBackend.flush();
+    });
+  });
+
+  describe("With Url", function() {
+    it("Shouldn't add suffix to URL", function() {
+      var suffixRestangular = Restangular.withConfig(function(RestangularConfigurer) {
+        RestangularConfigurer.setRequestSuffix('.json');
+      });
+
+      $httpBackend.expectGET('http://accounts.com/all');
+      suffixRestangular.allUrl('accounts', 'http://accounts.com/all').getList();
+      $httpBackend.flush();
+    });
+  });
+
+  describe("JSONp", function() {
+    it("should work for get", function() {
+      Restangular.setJsonp(true);
+      Restangular.one('accounts', 1).get();
+
+      $httpBackend.expectJSONP('/accounts/1');
+      $httpBackend.flush();
+    });
+
+    it("should work for getList", function() {
+      Restangular.setJsonp(true);
+      Restangular.all('accounts').getList();
+
+      $httpBackend.expectJSONP('/accounts');
+      $httpBackend.flush();
+    });
+
+    it("shouldn't override post", function() {
+      Restangular.setJsonp(true);
+      restangularAccounts.post({id: 2, user: "Someone"});
+
+      $httpBackend.expectPOST('/accounts').respond(201, '');
+      $httpBackend.flush();
+    });
+
+    
+  });
+
+  describe("Local data", function() {
+    it("Should restangularize a collection OK", function() {
+      var collection = angular.copy(accountsModel);
+
+      Restangular.restangularizeCollection(null, collection, 'accounts');
+
+      expect(_.has(collection, 'get')).toBe(true);
+      expect(_.has(collection[0], 'get')).toBe(true);
+
+      expect(collection.getRestangularUrl()).toBe('/accounts');
+      expect(collection[0].getRestangularUrl()).toBe('/accounts/0');
+
+    });
+  });
+
+  describe("$object", function() {
+    it("Should work for single get", function() {
+      var promise = Restangular.one('accounts', 1).get();
+      var obj = promise.$object;
+      expect(obj).toBeDefined();
+      expect(obj.amount).toBeUndefined();
+
+      $httpBackend.flush();
+
+      expect(obj.amount).toEqual(3.1416);      
+    });
+
+    it("Should be restangularized by default", function() {
+      Restangular.extendModel('accounts', function(account) {
+        account.extended = function() {return true;}
+        return account;
+      });
+
+      var promise = Restangular.one('accounts', 1).get();
+      var obj = promise.$object;
+      expect(obj).toBeDefined();
+      expect(obj.extended).toBeDefined();
+
+      $httpBackend.flush();
+    });
+
+    it("Should work for single get", function() {
+      var promise = Restangular.all('accounts').getList();
+      var list = promise.$object;
+      expect(list).toBeDefined();
+      expect(list.length).toEqual(0);
+
+      $httpBackend.flush();
+
+      expect(list.length).toEqual(2);
+      expect(list[1].amount).toEqual(3.1416);
+    });
+  });
+
   describe("ALL", function() {
     it("getList() should return an array of items", function() {
       restangularAccounts.getList().then(function(accounts) {
-        expect(sanitizeRestangularAll(accounts)).toEqual(sanitizeRestangularAll(accountsModel));
+        expect(Restangular.stripRestangular(accounts)).toEqual(Restangular.stripRestangular(accountsModel));
+      });
+
+      $httpBackend.flush();
+    });
+
+    it("several getList() should return an array of items", function() {
+      $httpBackend.expectGET('/accounts/0,1');
+      Restangular.several("accounts", 0, 1).getList().then(function(accounts) {
+        expect(Restangular.stripRestangular(accounts)).toEqual(Restangular.stripRestangular(accountsModel));
+      });
+
+      $httpBackend.flush();
+    });
+
+    it("several remove() should work", function() {
+      $httpBackend.expectDELETE('/accounts/0,1').respond([200, "", ""]);
+      Restangular.several("accounts", 0, 1).remove();
+
+      $httpBackend.flush();
+    });
+
+    it("get(id) should return the item with given id", function() {
+      restangularAccounts.get(0).then(function(account) {
+        expect(Restangular.stripRestangular(account)).toEqual(Restangular.stripRestangular(accountsModel[0]));
       });
 
       $httpBackend.flush();
@@ -102,9 +296,9 @@ describe("Restangular", function() {
       $httpBackend.flush();
     });
 
-    it("Custom GET methods sohuld work", function() {
+    it("Custom GET methods should work", function() {
       restangularAccounts.customGETLIST("messages").then(function(msgs) {
-        expect(sanitizeRestangularAll(msgs)).toEqual(sanitizeRestangularAll(messages));
+        expect(Restangular.stripRestangular(msgs)).toEqual(Restangular.stripRestangular(messages));
       });
 
       $httpBackend.flush();
@@ -119,6 +313,18 @@ describe("Restangular", function() {
     $httpBackend.flush();
    });
 
+    it("post() should work with arrays", function() {
+     Restangular.all('places').post([{name: "Gonto"}, {name: 'John'}]).then(function(value) {
+       expect(value.length).toEqual(2);
+     });
+
+    $httpBackend.expectPOST('/places').respond(function(method, url, data, headers) {
+      return [201, angular.fromJson(data), ""];
+    });
+
+    $httpBackend.flush();
+   });
+
     it("post() should add a new item with data and return the data from the server", function() {
      restangularAccounts.post(newAccount).then(function(added) {
        expect(added.fromServer).toEqual(true);
@@ -130,22 +336,19 @@ describe("Restangular", function() {
    });
 
     it("Doing a post and then other operation (delete) should call right URLs", function() {
-      restangularAccounts.getList().then(function(accounts) {
-        accounts[1].post('transactions', {id: 1, name: "Gonto"}).then(function(transaction) {
-          transaction.remove();
-          $httpBackend.expectDELETE('/accounts/1/transactions/1').respond(201, '');
-        });
-      });
-
+      restangularAccounts.post(newAccount).then(function(added) {
+        added.remove();
+        $httpBackend.expectDELETE('/accounts/44').respond(201, '');   
+      });      
+      
       $httpBackend.flush();
     });
 
-    it("Doing a post to a server that returns no element will return the parameter of that post", function() {
+    it("Doing a post to a server that returns no element will return undefined", function() {
       restangularAccounts.getList().then(function(accounts) {
         var newTransaction = {id: 1, name: "Gonto"};
         accounts[1].post('transactions', newTransaction).then(function(transaction) {
-          expect(sanitizeRestangularOne(transaction))
-            .toEqual(sanitizeRestangularOne(newTransaction));
+          expect(transaction).toBeUndefined();
         });
       });
 
@@ -167,10 +370,32 @@ describe("Restangular", function() {
       $httpBackend.flush();
     });
 
+    it("customPUT should work", function() {
+      $httpBackend.expectPUT('/accounts/hey').respond(accountsModel);
+      restangularAccounts.customPUT({key: 'value'}, 'hey');
+
+      $httpBackend.flush();
+    });
+
     it("options()  should safely return", function() {
       restangularAccounts.options().then(function() {
         expect(true).toBe(true);
       });
+
+      $httpBackend.flush();
+    });
+
+     it("getList() should correctly handle params after customDELETE", function() {
+      $httpBackend.expectGET('/accounts?foo=1').respond(accountsModel);
+      restangularAccounts.getList({foo: 1}).then(function(){
+        $httpBackend.expectDELETE('/accounts?id=1').respond(201, '');
+        return restangularAccounts.customDELETE('', {id: 1});
+      }).then(function() {
+          $httpBackend.expectGET('/accounts?foo=1').respond(accountsModel);
+          return restangularAccounts.getList({foo: 1});
+        }).then(function(accounts) {
+          expect(Restangular.stripRestangular(accounts)).toEqual(Restangular.stripRestangular(accountsModel));
+        });
 
       $httpBackend.flush();
     });
@@ -179,8 +404,8 @@ describe("Restangular", function() {
   describe("ONE", function() {
     it("get() should return a JSON item", function() {
       restangularAccount1.get().then(function(account) {
-        expect(sanitizeRestangularOne(account))
-          .toEqual(sanitizeRestangularOne(accountsModel[1]));
+        expect(Restangular.stripRestangular(account))
+          .toEqual(Restangular.stripRestangular(accountsModel[1]));
       });
 
       $httpBackend.flush();
@@ -188,8 +413,8 @@ describe("Restangular", function() {
 
     it("Should make RequestLess connections with one", function() {
       restangularAccount1.one("transactions", 1).get().then(function(transaction) {
-        expect(sanitizeRestangularOne(transaction))
-          .toEqual(sanitizeRestangularOne(accountsModel[1].transactions[1]));
+        expect(Restangular.stripRestangular(transaction))
+          .toEqual(Restangular.stripRestangular(accountsModel[1].transactions[1]));
       });
 
       $httpBackend.flush();
@@ -197,8 +422,8 @@ describe("Restangular", function() {
 
     it("Should make RequestLess connections with all", function() {
       restangularAccount1.all("transactions").getList().then(function(transactions) {
-        expect(sanitizeRestangularAll(transactions))
-          .toEqual(sanitizeRestangularAll(accountsModel[1].transactions));
+        expect(Restangular.stripRestangular(transactions))
+          .toEqual(Restangular.stripRestangular(accountsModel[1].transactions));
       });
 
       $httpBackend.flush();
@@ -207,7 +432,7 @@ describe("Restangular", function() {
 
     it("Custom GET methods should work", function() {
       restangularAccount1.customGET("message").then(function(msg) {
-        expect(sanitizeRestangularOne(msg)).toEqual(sanitizeRestangularOne(messages[0]));
+        expect(Restangular.stripRestangular(msg)).toEqual(Restangular.stripRestangular(messages[0]));
       });
 
       $httpBackend.flush();
@@ -232,8 +457,8 @@ describe("Restangular", function() {
     it("should return an array when accessing a subvalue", function() {
       restangularAccount1.get().then(function(account) {
         account.getList("transactions").then(function(transactions) {
-          expect(sanitizeRestangularAll(transactions))
-            .toEqual(sanitizeRestangularAll(accountsModel[1].transactions));
+          expect(Restangular.stripRestangular(transactions))
+            .toEqual(Restangular.stripRestangular(accountsModel[1].transactions));
         });
       });
 
@@ -241,4 +466,199 @@ describe("Restangular", function() {
     });
   });
 
+  describe("COPY", function() {
+    it("should copy an object and 'this' should reference the copied object", function() {
+      var copiedAccount = Restangular.copy(accountsModel[0]);
+      var that;
+
+      copiedAccount.user = "Copied string";
+      expect(copiedAccount).not.toBe(accountsModel[0]);
+
+      // create a spy for one of the methods to capture the value of 'this'
+      spyOn(copiedAccount, 'getRestangularUrl').andCallFake(function() {
+        that = this;
+      });
+
+      copiedAccount.getRestangularUrl(); // invoke the method we are spying on
+      expect(that).toBe(copiedAccount);
+    });
+  });
+
+  describe("getRestangularUrl", function() {
+    it("should return the generated URL when you chain Restangular methods together", function() {
+      var restangularSpaces = Restangular.one("accounts",123).one("buildings", 456).all("spaces");
+      expect(restangularSpaces.getRestangularUrl()).toEqual("/accounts/123/buildings/456/spaces");
+    });
+  });
+
+  describe("getRestangularUrl with useCannonicalId set to true", function() {
+    it("should return the generated URL when you chain Restangular methods together", function() {
+      var R = Restangular.withConfig(function(config) {
+        config.setUseCannonicalId(true);
+      });
+      var restangularSpaces = R.one("accounts",123).one("buildings", 456).all("spaces");
+      expect(restangularSpaces.getRestangularUrl()).toEqual("/accounts/123/buildings/456/spaces");
+    });
+  });
+
+
+  describe("addElementTransformer", function() {
+    it("should allow for a custom method to be placed at the collection level", function() {
+      var accountsPromise;
+
+      Restangular.addElementTransformer('accounts', true, function(collection) {
+         collection.totalAmount = function() {};
+         return collection;
+      });
+
+      accountsPromise = Restangular.all('accounts').getList();
+      
+      accountsPromise.then(function(accounts) {
+        expect(typeof accounts.totalAmount).toEqual("function");
+      });
+
+      $httpBackend.flush();
+    });
+
+    it("should allow for a custom method to be placed at the model level when one model is requested", function() {
+      var accountPromise;
+      
+      Restangular.addElementTransformer('accounts', false, function(model) {
+         model.prettifyAmount = function() {};
+         return model;
+      });
+
+      accountPromise = Restangular.one('accounts', 1).get();
+      
+      accountPromise.then(function(account) {
+        expect(typeof account.prettifyAmount).toEqual("function");
+      });
+
+      $httpBackend.flush();
+    });
+
+    it("should allow for a custom method to be placed at the model level when several models are requested", function() {
+      var accountPromise;
+      
+      Restangular.addElementTransformer('accounts', false, function(model) {
+         model.prettifyAmount = function() {};
+         return model;
+      });
+
+      accountsPromise = Restangular.all('accounts', 1).getList();
+      
+      accountsPromise.then(function(accounts) {
+        accounts.forEach(function(account, index) {
+          expect(typeof account.prettifyAmount).toEqual("function");
+        });
+      });
+
+      $httpBackend.flush();
+    });
+  });
+
+  describe("extendCollection", function() {
+    it("should be an alias for a specific invocation of addElementTransformer", function() {
+      var spy = spyOn(Restangular, 'addElementTransformer');
+
+      var fn = function(collection) {
+        collection.totalAmount = function() {};
+        return collection;
+      };
+
+      Restangular.extendCollection('accounts', fn);
+
+      expect(spy).toHaveBeenCalledWith('accounts', true, fn);
+    });
+  });
+
+  describe("extendModel", function() {
+    it("should be an alias for a specific invocation of addElementTransformer", function() {
+      var spy = spyOn(Restangular, 'addElementTransformer');
+
+      var fn = function(model) {
+        model.prettifyAmount = function() {};
+        return model;
+      };
+
+      Restangular.extendModel('accounts', fn);
+
+      expect(spy).toHaveBeenCalledWith('accounts', false, fn);
+    });
+  });
+  
+  describe("defaultHeaders", function() {
+    it("should return defaultHeaders", function() {
+      var defaultHeaders = {testheader:'header value'};
+      
+      Restangular.setDefaultHeaders(defaultHeaders);
+      
+      expect(Restangular.defaultHeaders).toEqual(defaultHeaders);
+    });
+  });
+  
+  describe("defaultRequestParams", function() {
+    it("should return defaultRequestParams", function() {
+      var defaultRequestParams = {param:'value'};
+      
+      Restangular.setDefaultRequestParams(defaultRequestParams);
+      
+      expect(Restangular.requestParams.common).toEqual(defaultRequestParams);
+    });
+    
+    it("should be able to set default params for get, post, put.. methods separately", function() {
+      var postParams = {post:'value'},
+          putParams = {put:'value'};
+      
+      Restangular.setDefaultRequestParams('post', postParams);
+      expect(Restangular.requestParams.post).toEqual(postParams);
+      
+      Restangular.setDefaultRequestParams('put', putParams);
+      expect(Restangular.requestParams.put).toEqual(putParams);
+      
+      expect(Restangular.requestParams.common).not.toEqual(putParams);
+    });
+    
+    it("should be able to set default params for multiple methods with array", function() {
+      var defaultParams = {param:'value'};
+      
+      Restangular.setDefaultRequestParams(['post', 'put'], defaultParams);
+      
+      expect(Restangular.requestParams.post).toEqual(defaultParams);
+      expect(Restangular.requestParams.put).toEqual(defaultParams);
+      
+      expect(Restangular.requestParams.common).not.toEqual(defaultParams);
+    });
+  });
+
+  describe("withConfig", function() {
+    it("should create new service with scoped configuration", function() {
+      var childRestangular = Restangular.withConfig(function(RestangularConfigurer){
+        RestangularConfigurer.setBaseUrl('/api/v1');
+      });
+
+      expect(Restangular.configuration.baseUrl).toEqual('');
+      expect(childRestangular.configuration.baseUrl).toEqual('/api/v1');
+      
+    });
+
+    it("should allow nested configurations", function() {
+      var childRestangular = Restangular.withConfig(function(RestangularConfigurer){
+        RestangularConfigurer.setBaseUrl('/api/v1');
+      });
+    
+      var grandchildRestangular = childRestangular.withConfig(function(RestangularConfigurer){
+        RestangularConfigurer.setRequestSuffix('.json');
+      });
+
+      expect(Restangular.configuration.baseUrl).toEqual('');
+      expect(Restangular.configuration.suffix).toEqual(null);
+
+      expect(childRestangular.configuration.baseUrl).toEqual('/api/v1');
+      expect(childRestangular.configuration.suffix).toEqual(null);
+
+      expect(grandchildRestangular.configuration.baseUrl).toEqual('/api/v1');
+      expect(grandchildRestangular.configuration.suffix).toEqual('.json');
+    });
+  });
 });