+++ www/components/restangular/CHANGELOG.md
@@ -1,3 +1,22 @@
+# New release information
+New releases are in Releases section of Github
+https://github.com/mgonto/restangular/releases
+
+#1.0.11
+* Documentation Typo fixes
+* errorInterceptor can now stop Restangular from rejecting the promise
+* Bugfix fot method override on DELETE. Now it works
+
+#1.0.9
+* **BREAKING CHANGE**: Restangular methods created with `addRestangularMethod` will change its signature depending on the opreation. If the operation is safe (GET, OPTIONS, etc.), the signature is methodName(params, headers, elemForBody). If it's not safe (POST, PUT, etc.), the signature is methodName(elemForBody, params, headers). This is to facilitate using them as when it's not safe, you're usually going to set a body
+* Now you can configure default request parameters per method and for everything as well
+* Added the ability to use Cannonical IDs. They're used if you need to change Primary Key (ID) of the element (Really weird case).
+* If response is null or undefined, the element sent in the request ISN'T used anymore. This is to have clarity of what's returned by the server and also to fix one bug.
+* Added tests
+* Fixed bug with ID when it was an empty string
+* Added missing ';'.
+
+
 #1.0.7
 * `baseUrl` can now be set either with or without ending `/` and it'll work